Fix rollup hanging while bundling caused by react directive and sourcemaps
