+++ packages/pluggable-widgets-tools/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We fixed an issue where the rollup process would sometimes hang and prevent the widget build/release script from completing. 
+
 ## [10.7.2] - 2024-03-06
 
 ### Fixed
