+++ packages/pluggable-widgets-tools/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mendix/pluggable-widgets-tools",
-  "version": "10.7.2",
+  "version": "10.7.3",
   "description": "Mendix Pluggable Widgets Tools",
   "engines": {
     "node": ">=16"
